Add support for syncing ADS
Add command --copy-cifsads.
Copies a hardcoded list of cifs filestreams.

Implies and mutually exclusive with --copy-cifsacls.

To simplify the code, rsync always includes the ADS name with the source
file info, but when source file has no ADS, the ADS data will be empty
and in this case rsync tries to remove ADS from the destination file.

Not sure if this is in SMB spec, but Samba rejects open of filestream
if any of the other filestreams (including main DATA) is already open.

Rsync does:
- open(file)
- read_and_send_xattr_and_ads()
- read_and_send_file_datasum()

It is not a problem to do the open() after reading the ads/xattr.
This does not need to depend on --copy-cifsads, but to reduce risk
make this change only with --copy-cifsads.

--copy-cifsads also requires cifs mount with -o actimeo=0, otherwise
the streams are kept open for one extra second and rsync fails.