Add support for syncing cifs ACLs
cifs exports ACLs with magic xattrs that are not listed in listxattr.
With --cifacls, sync the magic cifs ACL xattrs.

Because cifs client does not cache the cifs ACLs, getxattr() for
comparing cifs ACLs is expensive, so with --copy-cifsacls, force copy
the ACLs from source to dest file without comparing them first.

setxattr of cifs acls does not change mtime so no need to reset mtime
after it.

The --cifsacls options imply --xattrs.